abstract FieldValue

+++ cool-core/src/test/java/com/nus/cool/core/io/storevector/ZIntBitInputVectorTest.java
@@ -32,7 +32,7 @@ public class ZIntBitInputVectorTest {
     ZIntBitInputVector in = ZIntBitInputVector.load(buffer);
 
     for (int i = 0; i < in.size(); i++) {
-      Assert.assertEquals(in.get(i), numbers[i]);
+      Assert.assertEquals(in.get(i).intValue(), numbers[i]);
     }
 
   }
